+++ CollectionKitTests/FlowLayoutSpec.swift
@@ -22,40 +22,28 @@ class FlowLayoutSpec: QuickSpec {
         expect(layout.frames.count).to(equal(1))
       }
 
-      it("should work with both axis") {
+      it("should wrap child") {
         let layout = FlowLayout<CGSize>()
         layout.mockLayout(parentSize: (100, 100), (50, 50), (50, 50), (50, 50))
         expect(layout.frames).to(equal(frames((0, 0, 50, 50), (50, 0, 50, 50), (0, 50, 50, 50))))
-        layout.axis = .horizontal
-        layout.mockLayout(parentSize: (100, 100), (50, 50), (50, 50), (50, 50))
-        expect(layout.frames).to(equal(frames((0, 0, 50, 50), (0, 50, 50, 50), (50, 0, 50, 50))))
       }
 
-      it("should work with both axis with linespacing") {
+      it("should work with linespacing") {
         let layout = FlowLayout<CGSize>(lineSpacing: 10)
         layout.mockLayout(parentSize: (100, 300), (100, 100), (100, 100), (100, 100))
         expect(layout.frames).to(equal(frames((0, 0, 100, 100), (0, 110, 100, 100), (0, 220, 100, 100))))
-        layout.axis = .horizontal
-        layout.mockLayout(parentSize: (300, 100), (100, 100), (100, 100), (100, 100))
-        expect(layout.frames).to(equal(frames((0, 0, 100, 100), (110, 0, 100, 100), (220, 0, 100, 100))))
       }
 
-      it("should work with both axis with interitemspacing") {
+      it("should work with interitemspacing") {
         let layout = FlowLayout<CGSize>(interitemSpacing: 10)
         layout.mockLayout(parentSize: (130, 100), (50, 50), (50, 50), (50, 50))
         expect(layout.frames).to(equal(frames((0, 0, 50, 50), (60, 0, 50, 50), (0, 50, 50, 50))))
-        layout.axis = .horizontal
-        layout.mockLayout(parentSize: (100, 130), (50, 50), (50, 50), (50, 50))
-        expect(layout.frames).to(equal(frames((0, 0, 50, 50), (0, 60, 50, 50), (50, 0, 50, 50))))
       }
 
-      it("should work with both axis with interitemspacing and linespacing") {
+      it("should work with interitemspacing and linespacing") {
         let layout = FlowLayout<CGSize>(lineSpacing: 10, interitemSpacing: 10)
         layout.mockLayout(parentSize: (130, 130), (50, 50), (50, 50), (50, 50))
         expect(layout.frames).to(equal(frames((0, 0, 50, 50), (60, 0, 50, 50), (0, 60, 50, 50))))
-        layout.axis = .horizontal
-        layout.mockLayout(parentSize: (130, 130), (50, 50), (50, 50), (50, 50))
-        expect(layout.frames).to(equal(frames((0, 0, 50, 50), (0, 60, 50, 50), (60, 0, 50, 50))))
       }
     }
   }
