+++ CollectionKitTests/TestUils.swift
@@ -13,9 +13,9 @@ import UIKit.UIGestureRecognizerSubclass
 extension CollectionLayout where Data == CGSize {
 
   func mockLayout(parentSize: (CGFloat, CGFloat) = (300, 300), _ childSizes: (CGFloat, CGFloat)...) {
-    doLayout(collectionSize: CGSize(width: parentSize.0, height: parentSize.1),
-            dataProvider: ArrayDataProvider(data: sizes(childSizes)),
-            sizeProvider: { (index, data, collectionSize) -> CGSize in
+    layout(collectionSize: CGSize(width: parentSize.0, height: parentSize.1),
+           dataProvider: ArrayDataProvider(data: sizes(childSizes)),
+           sizeProvider: { (index, data, collectionSize) -> CGSize in
               return data
     })
   }
