+++ Sources/Addon/EmptyStateCollectionProvider.swift
@@ -19,7 +19,7 @@ public class EmptyStateCollectionProvider: CollectionComposer {
     self.emptyStateViewGetter = emptyStateView
     self.content = content
     super.init(identifier: identifier,
-               layout: InverseLayout(FlexLayout(flex: ["emptyStateView": FlexValue(flex: 1)])),
+               layout: InverseLayout(RowLayout("emptyStateView")),
                sections: [content])
   }
